Small code clean-up
